+++ zds/tutorialv2/models/database.py
@@ -37,7 +37,8 @@ from zds.tutorialv2.models.mixins import TemplatableContentModelMixin, OnlineLin
 from zds.tutorialv2.models.versioned import NotAPublicVersion
 from zds.tutorialv2.utils import get_content_from_json, BadManifestError, get_blob
 from zds.utils import get_current_user
-from zds.utils.models import SubCategory, Licence, HelpWriting, Comment, Tag
+from zds.utils.models import SubCategory, Licence, Comment, Tag
+from zds.tutorialv2.models.help_requests import HelpWriting
 from zds.utils.templatetags.emarkdown import render_markdown_stats
 from zds.utils.uuslug_wrapper import uuslug
