+++ zds/tutorialv2/tests/tests_views/tests_published.py
@@ -19,6 +19,7 @@ from zds.tutorialv2.tests.factories import (
     ContainerFactory,
     ExtractFactory,
     PublishedContentFactory,
+    HelpWritingFactory,
 )
 from zds.tutorialv2.models.database import (
     PublishableContent,
@@ -30,7 +31,7 @@ from zds.tutorialv2.models.database import (
 from zds.tutorialv2.publication_utils import publish_content
 from zds.tutorialv2.tests import TutorialTestMixin
 from zds.utils.models import Alert, Tag, Hat
-from zds.utils.tests.factories import CategoryFactory, SubCategoryFactory, LicenceFactory, HelpWritingFactory
+from zds.utils.tests.factories import CategoryFactory, SubCategoryFactory, LicenceFactory
 from zds.utils.header_notifications import get_header_notifications
 from copy import deepcopy
 from zds import json_handler
