+++ zds/utils/management/tests.py
@@ -4,7 +4,8 @@ from django.test import TestCase
 from django.contrib.auth.models import User, Permission
 from zds.member.models import Profile
 from zds.forum.models import Forum, Topic, ForumCategory
-from zds.utils.models import Tag, Category as TCategory, CategorySubCategory, SubCategory, HelpWriting, Licence
+from zds.utils.models import Tag, Category as TCategory, CategorySubCategory, SubCategory, Licence
+from zds.tutorialv2.models.help_requests import HelpWriting
 from zds.member.tests.factories import ProfileFactory
 from zds.tutorialv2.models.database import (
     PublishableContent,
