+++ zds/utils/models.py
@@ -15,8 +15,6 @@ from django.utils.translation import gettext_lazy as _
 from django.dispatch import receiver
 from django.template.loader import render_to_string
 
-from easy_thumbnails.fields import ThumbnailerImageField
-
 from zds.utils import signals
 from zds.mp.models import PrivateTopic
 from zds.tutorialv2.models import TYPE_CHOICES, TYPE_CHOICES_DICT
@@ -790,30 +788,3 @@ class Tag(models.Model):
 
     def has_object_read_permission(self, request):
         return True
-
-
-class HelpWriting(models.Model):
-
-    """Tutorial Help"""
-
-    class Meta:
-        verbose_name = "Aide à la rédaction"
-        verbose_name_plural = "Aides à la rédaction"
-
-    # A name for this help
-    title = models.CharField("Name", max_length=20, null=False)
-    slug = models.SlugField(max_length=20)
-
-    # tablelabel: Used for the accessibility "This tutoriel need help for writing"
-    tablelabel = models.CharField("TableLabel", max_length=150, null=False)
-
-    # The image to use to illustrate this role
-    image = ThumbnailerImageField(upload_to=image_path_help)
-
-    def __str__(self):
-        """Textual Help Form."""
-        return self.title
-
-    def save(self, *args, **kwargs):
-        self.slug = old_slugify(self.title)
-        super().save(*args, **kwargs)
